refactor: improve file organization in manager application
