+++ manager-application/src/app/item-dialog/item-dialog.component.ts
@@ -6,9 +6,9 @@ import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MAT_DIALOG_DATA, MatDialog, MatDialogContent } from '@angular/material/dialog';
-import { Ingredient } from '../../utils/ingredient';
+import { Ingredient } from '../../utils/schema/ingredient';
 import { IArray } from '../../utils/array';
-import { MenuServiceMessages, RequestMessage, ResponseMessage, WarehouseServiceMessages } from '../../utils/messages';
+import { MenuServiceMessages, RequestMessage, ResponseMessage, WarehouseServiceMessages } from '../../utils/schema/messages';
 import { Service } from '../../utils/service';
 import { checkWsConnectionAndSend } from '../../utils/send';
 import { buildRecipe } from '../../utils/recipe';
