+++ manager-application/src/utils/schema/item.ts
